Cleaned out unsorted files, organized condensed files for structures and notes
