--- Unsorted/multiple-choice-clarification.js
@@ -1,18 +1,0 @@
-function kthNode(node, k) {
-    let curr = node;
-    for (let i = 0; i < k; i++) {
-        curr = cujrr.next;
-    };
-
-    return curr;
-};
-
-/*
-1   2   3   4
-
-if k = 2 return 2.
-if k = 4 return 4
-if k = 5 return null
-*/
-
-
